requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
paddlepaddle==2.5.2
paddleocr==2.7.0.3
opencv-python==4.8.1.78
matplotlib==3.8.0
pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
